+++ src/Triangulo.cpp
@@ -1,13 +1,13 @@
 #include "../include/Triangulo.hpp"
 #include <iostream>
 
-Triangulo::Triangulo(int _matriz[3][3])
+Triangulo::Triangulo(float _matriz[3][3])
 {
     this->setMatriz(_matriz);
     this->resultado = sentidoPositivo() - sentidoNegativo();
 }
 
-void Triangulo::setMatriz(int _matriz[3][3]) {
+void Triangulo::setMatriz(float _matriz[3][3]) {
 
     for (int i = 0; i < 3; ++i) {
         for (int j = 0; j < 3; ++j) {
@@ -16,8 +16,8 @@ void Triangulo::setMatriz(int _matriz[3][3]) {
     }
 }
 
-int Triangulo::sentidoPositivo() {
-    int somatorio = 0;
+float Triangulo::sentidoPositivo() {
+    float somatorio = 0;
 
     somatorio+= m[0][0] * m[1][1] * m[2][2];
     somatorio+= m[0][1] * m[1][2] * m[2][0];
@@ -26,7 +26,7 @@ int Triangulo::sentidoPositivo() {
     return somatorio;
 }
 
-int Triangulo::sentidoNegativo() {
+float Triangulo::sentidoNegativo() {
      int somatorio = 0;
 
     somatorio+= m[0][2] * m[1][1] * m[2][0];
@@ -36,7 +36,7 @@ int Triangulo::sentidoNegativo() {
     return somatorio;
 }
 
-int Triangulo::getResultado() const {
+float Triangulo::getResultado() const {
     return resultado;
 }
